README for this Web Scraper
This is from Thiago Marzagao's blog http://thiagomarzagao.com/
I've changed a few things, but otherwise there isnt too much that needs changing.  I commented it so its a bit easier to follow.  Also, I chose to use firefox as my browser, but it is possible to use others if need be.